+++ src/utils/devicegeneric.hpp
@@ -71,8 +71,8 @@ public:
 	virtual int getLongValue(std::string attr);
 	virtual int getLongValue(unsigned int, std::string attr, bool output=false);
 	virtual int getBufferLongValue(std::string attr);
-	virtual int setLongValue(int value, std::string attr);
-	virtual int setLongValue(unsigned int chn_idx, int value, std::string attr, bool output=false);
+	virtual int setLongValue(long long value, std::string attr);
+	virtual int setLongValue(unsigned int chn_idx, long long value, std::string attr, bool output=false);
 	virtual int setBufferLongValue(int value, std::string attr);
 
 	virtual bool getBoolValue(std::string attr);
